fix shutdown crash
The issue previously was that the old function was just *requesting* shutdown, but we were still running in main frame loop, so the function that we hooked originally had itself main frame hook in the call chain. Meaning that freeing SDK and undoing hooks right there would still have us return to the hooked main loop func, and it'd try to continue executing with free'd SDK object, which would not end well. :)